requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
et-xmlfile==1.1.0
idna==3.2
openpyxl==3.0.7
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.6
